iss230: Forfeit Button
I moved the button on the multi and single to the bottom right of the scene and made sure that both look cohesive. I also moved the multi score board to the top left and made sure it is the same size as the single score board(which is on the top left).